requirements.txt
mcp
google-genai
python-dotenv
tavily-python
pypdf
txtai
torch